Added wrapper for numpy and cython implementations.  Includes timing within functinos.
A fair amount of work needs to be done for cython.  Note the commented
out functions.

Cython with c++ and AVX is a horror.  AVX does not play well with the
c++ compilier, so anything that uses AVX MUST be separate from the
c++ code.  This might lead to issues (we'll see).